requirements.txt
# Core Django dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Task Queue & Caching
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-redis==5.4.0
hiredis==2.2.3

# Payment Processing
stripe==7.8.0

# File Handling & Media
# Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0

# Notifications
twilio==8.10.3

# QR Code Generation
qrcode[pil]==7.4.2

# PDF Generation
reportlab==4.0.7

# Environment Variables
python-decouple==3.8

# Development & Testing
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2

# API Documentation
drf-spectacular==0.26.5

# Filtering
django-filter==23.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3